Rename KotlinxSerializers package
