+++ shared/src/commonMain/kotlin/org/itmodreamteam/myrest/shared/user/ActiveSession.kt
@@ -2,7 +2,7 @@ package org.itmodreamteam.myrest.shared.user
 
 import kotlinx.datetime.LocalDateTime
 import kotlinx.serialization.Serializable
-import org.itmodreamteam.myrest.shared.config.LocalDateTimeSerializer
+import org.itmodreamteam.myrest.shared.config.kotlinx.LocalDateTimeSerializer
 
 @Serializable
 data class ActiveSession(
